Changed moving platform file name
