+++ src/movingPlatform.js
